@@ -1,6 +1,5 @@
-﻿namespace RobotAndMaze.Domain.Models.Abstractions
+﻿namespace RobotAndMaze.Domain.Models.Abstractions;
+
+public interface IHelicopter : IRobot, IFly, IMove
 {
-    public interface IHelicopter : IRobot, IFly, IMove
-    {
-    }
 }
