@@ -1,10 +1,9 @@
 ﻿using RobotAndMaze.Domain.Models;
 using RobotAndMaze.Support;
 
-namespace RobotAndMaze.Domain.Strategies
+namespace RobotAndMaze.Domain.Strategies;
+
+public interface IMoveStrategy
 {
-    public interface IMoveStrategy
-    {
-        Result<Coordinates> CanMove(Matrix matrix, Direction direction, RobotType robotType);
-    }
+    Result<Coordinates> CanMove(Matrix matrix, Direction direction, RobotType robotType);
 }
