@@ -1,16 +1,15 @@
 ﻿using RobotAndMaze.Domain.Models;
 using RobotAndMaze.Support;
 
-namespace RobotAndMaze.Infrastructure
+namespace RobotAndMaze.Infrastructure;
+
+public class MatrixProvider : IMatrixProvider
 {
-    public class MatrixProvider : IMatrixProvider
+    public Matrix GetBasic()
     {
-        public Matrix GetBasic()
+        return new(MatrixSeed.Create6X6Cells(), new Coordinates
         {
-            return new(MatrixSeed.Create6X6Cells(), new Coordinates
-            {
-                XPos = 0, YPos = 0
-            });
-        }
+            XPos = 0, YPos = 0
+        });
     }
 }
