@@ -1,30 +1,29 @@
 ﻿using System;
 
-namespace RobotAndMaze.Support
+namespace RobotAndMaze.Support;
+
+public sealed class Result<T> : Result
 {
-    public sealed class Result<T> : Result
+    private readonly T value;
+
+    public Result(
+        T value,
+        bool success,
+        string error) : base(success, error)
     {
-        private readonly T value;
+        this.value = value;
+    }
 
-        public Result(
-            T value,
-            bool success,
-            string error) : base(success, error)
+    public T Value
+    {
+        get
         {
-            this.value = value;
-        }
-
-        public T Value
-        {
-            get
+            if (!this.Success)
             {
-                if (!this.Success)
-                {
-                    throw new InvalidOperationException("The result object has no value.");
-                }
-
-                return this.value;
+                throw new InvalidOperationException("The result object has no value.");
             }
+
+            return this.value;
         }
     }
 }
